SpringSecurity Implementation for JournalEntry

+++ src/main/java/net/engineeringdigest/journalApp/Services/JournalServices.java
@@ -9,55 +9,74 @@ import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.Date;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Component
 public class JournalServices {
 
     @Autowired
     private JournalEntryRepo journalEntryRepo;
+    @Autowired
+    private UserEntryRepo userEntryRepo;
+
 
     @Autowired
     private UserEntryRepo userDB;
 
     @Transactional
-    public ResponseEntity<?> createJournalEntriesOfUser(JournalEntry journalEntry, String username) {
+    public ResponseEntity<?> createJournalEntriesOfUser(JournalEntry journalEntry) {
         try {
             Date date = new Date();
             journalEntry.setDate(date);
-            Users user = userDB.findByusername(username);
+            Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+            Users userIndb = userEntryRepo.findByusername(auth.getName());
+            Users user = userDB.findByusername(userIndb.getUsername());
             JournalEntry saved = journalEntryRepo.save(journalEntry);
             user.getJournalEntries().add(saved);
             userDB.save(user);
-            return new ResponseEntity<>(user,HttpStatus.CREATED);
+            return new ResponseEntity<>(user, HttpStatus.CREATED);
         } catch (Exception e) {
-            return new ResponseEntity<>(e,HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(e, HttpStatus.BAD_REQUEST);
         }
-
     }
 
-    public List<JournalEntry> getEntry(String username) {
-        Users user = userDB.findByusername(username);
+    public List<JournalEntry> getEntry() {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Users userIndb = userEntryRepo.findByusername(auth.getName());
+        Users user = userDB.findByusername(userIndb.getUsername());
         return user.getJournalEntries();
     }
 
-    public Optional<JournalEntry> getEntryID(ObjectId id) {
-        return journalEntryRepo.findById(id);
+    public ResponseEntity<?> getEntryID(ObjectId id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Users userIndb = userEntryRepo.findByusername(auth.getName());
+        List<JournalEntry> list = userIndb.getJournalEntries().stream().filter(x -> x.getId().equals(id)).collect(Collectors.toList());
+        if (!list.isEmpty()) {
+            return new ResponseEntity<>(journalEntryRepo.findById(id).orElse(null), HttpStatus.FOUND);
+        } else {
+            return new ResponseEntity<>("Given ID is not associated With your Username or It may not be Present.", HttpStatus.NOT_FOUND);
+        }
     }
 
     public ResponseEntity<?> UpdateData(JournalEntry newEntry, ObjectId JournalID) {
-        JournalEntry oldEntry = journalEntryRepo.findById(JournalID).orElse(null);
-        if (oldEntry != null) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Users userIndb = userEntryRepo.findByusername(auth.getName());
+        List<JournalEntry> list = userIndb.getJournalEntries().stream().filter(x -> x.getId().equals(JournalID)).collect(Collectors.toList());
+        if (!list.isEmpty()) {
+            JournalEntry oldEntry = list.get(0);
             oldEntry.setTitle(newEntry.getTitle());
             oldEntry.setContent(newEntry.getContent());
             return new ResponseEntity<>(journalEntryRepo.save(oldEntry), HttpStatus.CREATED);
         }
-        return new ResponseEntity<>(HttpStatus.NOT_MODIFIED);
+        return new ResponseEntity<>("Given ID is not associated With your Username or It may not be Present.", HttpStatus.FORBIDDEN);
     }
 
     public ResponseEntity<?> getJournalID(String username) {
@@ -66,11 +85,16 @@ public class JournalServices {
         return new ResponseEntity<>(list, HttpStatus.FOUND);
     }
 
-    public ResponseEntity<?> deleteJournalEntry(String username, ObjectId journalID) {
-        Users user = userDB.findByusername(username);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(journalID));
-        userDB.save(user);
-        journalEntryRepo.deleteById(journalID);
-        return new ResponseEntity<>(HttpStatus.OK);
+    public ResponseEntity<?> deleteJournalEntry( ObjectId journalID) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+        Users userIndb = userEntryRepo.findByusername(auth.getName());
+        List<JournalEntry> list = userIndb.getJournalEntries().stream().filter(x -> x.getId().equals(journalID)).collect(Collectors.toList());
+        if (!list.isEmpty()) {
+            userIndb.getJournalEntries().removeIf(x -> x.getId().equals(journalID));
+            userDB.save(userIndb);
+            journalEntryRepo.deleteById(journalID);
+            return new ResponseEntity<>("Record Delete Successfully",HttpStatus.OK);
+        }
+        return new ResponseEntity<>("Given ID is not associated With your Username or It may not be Present.", HttpStatus.FORBIDDEN);
     }
 }
